% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONExtentRounder.R
\name{FUSIONExtentRounder}
\alias{FUSIONExtentRounder}
\title{FUSION extent rounder}
\usage{
FUSIONExtentRounder(ext, cellSize, snap = NULL, align = NULL,
  spill = "in")
}
\arguments{
\item{ext}{An \code{\link[raster]{Extent}} object.}

\item{cellSize}{numeric. The anticipated cell size of the raster that will be created by FUSION.}

\item{snap}{numeric. Vector of two numbers. An alternative to \code{align}, by which the output will snap to a pair of coordinates.}

\item{align}{character. Defines which corner of \code{ext} the output extent vector will be aligned. Accepts
\code{"topleft"}, \code{"topright"}, \code{"bottomleft"}, or \code{"bottomright"}.}

\item{spill}{If the coordinates of \code{ext} are not multiples of \code{cellSize}, the output will either be rounded inwards
when \code{spill} is set to \code{"in"}, or outwards if set to \code{"out"}.}
}
\value{
numeric. A vector of four numbers: X1, Y1, X2, Y2, which can be used as inputs for various \link[=FUSIONtools]{FUSION tools}.
}
\description{
This tool computes a vector to be used as the \code{gridxy} input for various \link[=FUSIONtools]{FUSION tools}
When establishing the extent of a raster, FUSION will align the CENTRE of the raster cells to the limits of the
"gridxy" vector, as opposed to the cells' edges. To compensate for this, this tool will create a vector that defines
an extent as being a half-cell size smaller in all directions than what is desired. In this way, the raster created
by FUSION will have cells whose edges align with the desired extent.
}
