% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalOptions.R
\name{FUSIONfolder}
\alias{FUSIONfolder}
\title{Set FUSION folder path}
\usage{
FUSIONfolder(path = NULL)
}
\arguments{
\item{path}{character. A path directing to the FUSION folder}
}
\value{
If no input is given for \code{path}, the current folder set for the \code{rFUSION.path}
global option is given.
}
\description{
A function to set the path to the FUSION directory. This directory contains the executable files employed by the
functions in rFUSION. The default location is \code{C:\\FUSION}. This path is stored as the \code{rFUSION.path}
global option.
}
