% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalOptions.R
\name{FUSIONprintToConsole}
\alias{FUSIONprintToConsole}
\title{Print to Console}
\usage{
FUSIONprintToConsole(value = NULL)
}
\arguments{
\item{value}{logical. FUSION tool outputs will print to console if set to \code{TRUE}}
}
\value{
If no input is given, the current value of the \code{rFUSION.printToConsole} global option is returned
}
\description{
A function for changing the \code{rFUSION.printToConsole} global option, which sets whether or not the FUSION tool
outputs are printed to the console. If no value is set, this function will return the current value.
}
