% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONTileClip.R
\name{FUSIONTileClip}
\alias{FUSIONTileClip}
\title{FUSION tile clip}
\usage{
FUSIONTileClip(inLAS, tiles, outFolder, prefix = NULL, ground = NULL,
  biaselev = NULL, zmin = NULL, zmax = NULL, verbose = TRUE)
}
\description{
Uses a tile scheme to clip a LAS dataset into tiles
}
