% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseTools.R
\name{FUSIONtools}
\alias{FUSIONtools}
\alias{ASCII2DTM}
\alias{CanopyModel}
\alias{ClipData}
\alias{CloudMetrics}
\alias{CSV2Grid}
\alias{DTM2ASCII}
\alias{DTM2TIF}
\alias{GridMetrics}
\alias{GridSurfaceCreate}
\alias{GroundFilter}
\alias{LDA2ASCII}
\alias{PolyClipData}
\title{FUSION tools}
\usage{
ASCII2DTM(surfacefile, xyunits = "m", zunits = "m", coordsys = 0,
  zone = 0, horizdatum = 0, vertdatum = 0, Gridfile, multiplier = NULL,
  offset = NULL)

CanopyModel(surfacefile, cellsize, datafile, xyunits = "m", zunits = "m",
  coordsys = 0, zone = 0, horizdatum = 0, vertdatum = 0,
  median = NULL, smooth = NULL, texture = NULL, slope = NULL,
  aspect = NULL, outlier = NULL, multiplier = NULL, class = NULL,
  ground = NULL, ascii = NULL, grid = NULL, gridxy = NULL,
  align = NULL, extent = NULL, surface = NULL, peaks = NULL,
  pointcount = NULL)

ClipData(InputSpecifier, SampleFile, MinX, MinY, MaxX, MaxY, shape = NULL,
  decimate = NULL, ground = NULL, zmin = NULL, zmax = NULL,
  zpercent = NULL, height = NULL, timemin = NULL, timemax = NULL,
  zero = NULL, biaselev = NULL, return = NULL, class = NULL,
  line = NULL, noindex = NULL, lndex = NULL, lda = NULL,
  nooffset = NULL, precision = NULL)

CloudMetrics(InputDataSpecifier, OutputFileName, above = NULL, new = NULL,
  firstinpulse = NULL, firstreturn = NULL, highpoint = NULL,
  subset = NULL, id = NULL, minht = NULL, maxht = NULL,
  outlier = NULL, strata = NULL, intstrata = NULL, kde = NULL,
  rgb = NULL)

CSV2Grid(inputfile, column, outputfile, multiplier = NULL, ndzero = NULL)

DTM2ASCII(inputfile, outputfile, csv = NULL, raster = TRUE,
  multiplier = NULL)

DTM2TIF(inputfile, outputfile = NULL, mask = NULL)

GridMetrics(groundfile, heightbreak, cellsize, outputfile, datafile,
  outlier = NULL, class = NULL, id = NULL, minpts = NULL,
  minht = NULL, nocsv = NULL, noground = NULL, nointdtm = NULL,
  diskground = NULL, alldensity = NULL, first = NULL, intensity = NULL,
  nointensity = NULL, rgb = NULL, fuel = NULL, grid = NULL,
  gridxy = NULL, align = NULL, extent = NULL, buffer = NULL,
  cellbuffer = NULL, strata = NULL, kde = NULL, ascii = NULL,
  topo = NULL, raster = NULL)

GridSurfaceCreate(surfacefile, cellsize, datafile, xyunits = "m",
  zunits = "m", coordsys = 0, zone = 0, horizdatum = 0, vertdatum = 0,
  median = NULL, smooth = NULL, slope = NULL, spike = NULL,
  gridxy = NULL, align = NULL, extent = NULL, residuals = NULL,
  class = NULL, minimum = NULL)

GroundFilter(outputfile, cellsize, datafile, surface = NULL, median = NULL,
  smooth = NULL, finalsmooth = NULL, outlier = NULL, gparam = NULL,
  wparam = NULL, aparam = NULL, bparam = NULL, tolerance = NULL,
  iterations = NULL, class = NULL, extent = NULL, trim = NULL,
  diagnostics = NULL, lda = NULL, precision = NULL)

LDA2ASCII(InputFile, OutputFile, format)

PolyClipData(Polyfile, Outputfile, Datafile, multifile = NULL, class = NULL)
}
\arguments{
\item{xyunits}{character. \code{"m"} for meters, \code{"f"} for feet.}

\item{zunits}{character. \code{"m"} for meters, \code{"f"} for feet.}

\item{coordsys}{numeric. Geographic coordinat system. 0 for unknown, 1 for UTM, 2 for state plane.}

\item{zone}{numeric. Coordinate system zone. 0 for unknown.}

\item{horizdatum}{numeric. Horizontal datum. 0 for unknown, 1 for NAD27, 2 for NAD83}

\item{vertdatum}{numeric. Vertical datum. 0 for unknown, 1 for 1 for NGVD29, 2 for NAVD88,3 for GRS80.}

\item{cellsize}{numeric. Cell size of an output raster file.}

\item{class}{numeric. Used to restrict a function to certain classified points.}

\item{ground}{A PLANS format DTM file representing a digital elevation model (DEM). Generally used to
normalize a point cloud to height above ground.}

\item{gridxy}{numeric. Vector of four numbers: X1, Y1, X2, Y2. Force the origin of an output raster (lower left corner)
to be (X1,Y1) instead of computing an origin from the data extents and force the upper right corner to be (X2, Y2).
X2 and Y2 will be rounded up to a multiple of \code{cellsize}. \strong{WARNING:} the \emph{center} of the outpur raster
cells will be aligned to the vector defined by \code{gridxy}, not the cell edges. To facilitate the use of \code{gridxy},
use the \code{\link{FUSIONExtentRounder}} function.}

\item{align}{A PLANS format DTM file to which an output raster will be aligned. Same behavior as \code{gridxy}, except
parameters are read from the DTM file.}

\item{extent}{A PLANS format DTM file. Force the origin and extent of the output grid to match the lower left corner
and extent of the specified DTM file but adjust the origin to be an even multiple of the cell size and the width and
height to be multiples of the cell size.}
}
\description{
Functions for calling FUSION tools. Consult the \href{http://forsys.cfr.washington.edu/fusion/FUSION_manual.pdf}{FUSION manual}
for more details on these tools, their inputs and their outputs. FUSION must be installed and the \code{rFUSION.path}
global option must direct to the FUSION folder in which all tool executable files are stored. The
\code{\link{FUSIONfolder}} function can be used to define this path.
}
\details{
FUSION is created and maintained by Robert J. McGaughey, of the U.S. Forest Service.
}
\section{Functions}{
\itemize{
\item \code{ASCII2DTM}: Convert an ASC file to a PLANS format DTM file

\item \code{CanopyModel}: Generate a Canopy Height Model (CHM) from a point cloud

\item \code{ClipData}: Clip a point cloud to a rectangular extent

\item \code{CloudMetrics}: Produce a CSV file of cloud metrics for a point cloud

\item \code{CSV2Grid}: Converts data stored in comma separated value (CSV) format into ASCII raster format

\item \code{DTM2ASCII}: Convert a PLANS format DTM file to an ASC file

\item \code{DTM2TIF}: Convert a PLANS format DTM file to a TIF file

\item \code{GridMetrics}: Computes a series of descriptive statistics for a LIDAR data set

\item \code{GridSurfaceCreate}: Create a gridded surface (i.e.: a raster) from a point cloud

\item \code{GroundFilter}: Filter out ground points for a point cloud

\item \code{LDA2ASCII}: Convert a LDA or LAS file to an ASCII text file

\item \code{PolyClipData}: Clip a point cloud to a shapefile (SHP)
}}

\references{
FUSION created by Robert J. McGaughey. U.S. Department of Agriculture, Forest
Service, Pacific Northwest Research Station, University of Washington, Box 352100,
Seattle, WA 98195-2100, \email{bmcgaughey@fs.fed.us}

FUSION website: \url{http://forsys.cfr.washington.edu/fusion/fusionlatest.html}
}
