% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricRaster.R
\name{MetricRaster}
\alias{MetricRaster}
\title{Metric Raster}
\usage{
MetricRaster(inCSV, outFolder, metrics, tileName = NULL, subHgtBreak = TRUE)
}
\arguments{
\item{inCSV}{character. Path for input CSV created from the \code{GridMetrics} function}

\item{outFolder}{character. Output folder}

\item{metrics}{character. Names of metrics as they appear in the \code{Abbrev} column in the \code{metricNames}
table}

\item{tileName}{character. Optional name for output files. If set to NULL, output raster files will be given the
name of the metric. If \code{tileName} is given a value, then each raster file will be put in a folder named
after its metric, and the file itself will be given the \code{tileName} value.}

\item{subHgtBreak}{logical. If FALSE, output files involving height breaks will use the placeholder "minHeight",
if set to TRUE, "minHeight" will be substituted by its corresponding numerical height break value.}
}
\description{
Convert grid metrics from a CSV file into multiple rasters
}
